Removed certificate interval for broker.
It is no longer required because for each accept, ssl::context is newly
created. So if certificate and key are updated, then it reflects from
the next accept.

NOTE: Even if timer is used, accepting certificate is not updated. It
reflects from the next accept. So this fix has no disadvantage.