Fixed disconnect status updating timing.
When DISCONNECT packet is received or invalid packet is received,
status_ shouldn't be updated to disconnected.
If send() is called just after that, sent packet is NOT stored
even if store is required.
It sometimes happens in the real use case.

So this fix removed status update to disconnect on recv().
After error is reported, the user call notify_closed(), then status is
updated to disconnect.